requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
scikit-learn
python-multipart
